#9P commit : Reorder the class
- create new class for splash_page
- also move widget(logo) class in to splash_page
- recreate default file (widget_test)
- make class main minimal
