+++ lib/login.dart
@@ -18,27 +18,28 @@ class LoginPageState extends State<LoginPage>
 
   final _formkey = GlobalKey<FormState>();
   ApiProvider apiProvider = ApiProvider();
-  Future doLogin() async {
-    if (_formkey.currentState.validate()) {
-      try {
-        var rs = await apiProvider.doLogin(_username.text, _password.text);
-        if (rs.statusCode == 200) {
-          print(rs.body);
-          var jsonRes = json.decode(rs.body);
-          if (jsonRes['ok']) {
-            String token = jsonRes['token'];
-            print(token);
-          } else {
-            print('Server error');
-          }
-        } else {
-          print('server error');
-        }
-      } catch (e) {
-        print(e);
-      }
-    }
-  }
+
+  // Future doLogin() async {
+  //   if (_formkey.currentState.validate()) {
+  //     try {
+  //       var rs = await apiProvider.doLogin(_username.text, _password.text);
+  //       if (rs.statusCode == 200) {
+  //         print(rs.body);
+  //         var jsonRes = json.decode(rs.body);
+  //         if (jsonRes['ok']) {
+  //           String token = jsonRes['token'];
+  //           print(token);
+  //         } else {
+  //           print('Server error');
+  //         }
+  //       } else {
+  //         print('server error');
+  //       }
+  //     } catch (e) {
+  //       print(e);
+  //     }
+  //   }
+  // }
 
   @override
   void initState() {
